Update Charts_BLS.py
Working on the makeHorzBarChart function. It's not general enough, and leave issues to be had in January. It also doesn't allow us to look at anything other than 1 month back and 12 months back. It also requires 12 series, but I want to make it so that I can have any number of series, even just one.